Move Api repository to Core
