+++ src/Wallabag/CoreBundle/Controller/Api/DeveloperController.php
@@ -9,9 +9,9 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Wallabag\ApiBundle\Form\Type\ClientType;
-use Wallabag\ApiBundle\Repository\ClientRepository;
 use Wallabag\CoreBundle\Controller\AbstractController;
 use Wallabag\CoreBundle\Entity\Api\Client;
+use Wallabag\CoreBundle\Repository\Api\ClientRepository;
 
 class DeveloperController extends AbstractController
 {
